libicuuc: Fix memory corruption in libicuuc
BZ:27414

This patch is used to replace patch-31604 ported from R2, which is not
a right fix to bz14409&bz22183 and still has potential mem leak/corruption
issue.

The ISCI routines are making use of some predefined sets which gets
allocated and initialized in the ISCI open routine and freed in the
ISCI close routine.

These predefined sets are global, so if we use ISCI operations with
multiple nested contexts we can easily run into memory corruption
issues.

An example for a double free issue:

_ISCIIOpen (creates context 1) -> PNJ_CONSONANT_SET = malloc(...)
_ISCIIOpen (creates context 2) -> PNJ_CONSONANT_SET = malloc(...)
...
_ISCIIClose (context 1) -> free(PNJ_CONSONANT_SET)
_ISCIIClose (context 2) -> free(PNJ_CONSONANT_SET) --> double free

If multiple threads are used then we can also run into use before
initialization errors beside double free errors.

To fix this issue we allocate and initialize the global sets only if
this is the first open and we free the global sets only if this is the
last close.

OTC-Bug: ANDROID-835

Change-Id: Ia4e3f4566c7d8c8b43fd22c0621f570c19163a27
Original-Change-Id: I042292a8413802f4b204ce0330c84a27e066e48a
OTC-Change-Id: I7876310b047ba5b8a15f1df62ea99bbcc473c884
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Du Chenyang <chenyang.du@intel.com>
Reviewed-on: http://android.intel.com:8080/79215
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>